+++ src/ratelimit.rs
@@ -134,7 +134,7 @@ impl Endpoint for RateLimitEndpoint {
 
     // fusa:req REQ-RL-005
     // fusa:req REQ-RL-006
-    fn read(&self, read_size: u8) -> Result<Vec<u8>, RcpError> {
+    fn read(&self, read_size: u16) -> Result<Vec<u8>, RcpError> {
         self.consume()?;
         self.inner.read(read_size)
     }
